+++ Sources/WebView/YouTubePlayerWebView+API.swift
@@ -500,7 +500,7 @@ extension YouTubePlayerWebView: YouTubePlayerPlaybackAPI {
     /// Retrieve the elapsed time in seconds since the video started playing
     /// - Parameter completion: The completion closure
     func getCurrentTime(
-        completion: @escaping (Result<Int, YouTubePlayerAPIError>) -> Void
+        completion: @escaping (Result<Double, YouTubePlayerAPIError>) -> Void
     ) {
         self.evaluate(
             javaScript: "player.getCurrentTime();",
@@ -532,7 +532,7 @@ extension YouTubePlayerWebView: YouTubePlayerVideoInformationAPI {
     /// Retrieve the duration in seconds of the currently playing video
     /// - Parameter completion: The completion closure
     func getDuration(
-        completion: @escaping (Result<Int, YouTubePlayerAPIError>) -> Void
+        completion: @escaping (Result<Double, YouTubePlayerAPIError>) -> Void
     ) {
         self.evaluate(
             javaScript: "player.getDuration();",
